+++ tests/run_pyflakes.sh
@@ -6,7 +6,9 @@ PYTHON_ROOT="${ROOT_DIR}"
 OUT_FILE="${ROOT_DIR}/tests/pyflakes.txt"
 
 cd "${PYTHON_ROOT}"
-pyflakes $(find "${PYTHON_ROOT}" -type f -name "*.py" -print) > "${OUT_FILE}" 2>&1 || true
+pyflakes $(find "${PYTHON_ROOT}" \
+	-type d -name "legacy" -prune -o \
+	-type f -name "*.py" -print) > "${OUT_FILE}" 2>&1 || true
 RESULT=$(wc -l < "${OUT_FILE}")
 if [ "${RESULT}" -eq 0 ]; then
 	echo "No errors found!!!"